pyproject.toml
[project]
name = "swarm-hr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "swarm",
    "black>=24.10.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
swarm = { git = "ssh://git@github.com/openai/swarm.git" }
